pyproject.toml
[project]
name = "repl-nix-financetrackpro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "components>=0.0.1a0",
    "gocardless-pro>=1.49.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.40.1",
]
